requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
websocket-client==1.7.0
python-dotenv==1.0.0
loguru==0.7.2